+++ projects/triangular-engine/jolt/vehicle/index.ts
@@ -1,0 +1,6 @@
+export * from './jolt-vehicle-constraint.component';
+export * from './jolt-vehicle-wheel.component';
+export * from './jolt-tracked-vehicle.component';
+export * from './jolt-wheeled-vehicle.component';
+export * from './jolt-tracked-wheel.component';
+export * from './jolt-wheeled-wheel.component';